Check config file and clean up
Made function to check if the contents of the config file had all of the required keys. It currently overwrites the file if it doesn't so this need to be changed to just add the non existing keys. Also removed unnecessary code and added options to change the names of some of the variables that it will be reading from the JSON file it gets from Manage Engine. Also found an error in the noticeboard.js; the get-data function was missing and two instances of populatedata.