Config file upgrade
App will check if config file exists and create it if it doesn't. If it does it will iterate over the required keys to check if all are available. If not it will only upgrade the missing keys without destroying existing config.

Also added the beginings of the pyproject.toml file to make the app installable. And removed some pointless code from javascript files.

Need to add a way for python to send the campus and status names to the javascript so it can draw them onto the html.